+++ docs/tutorials/portable-consent/_category_.json
@@ -1,5 +1,5 @@
 {
-  "label": "Integrations",
+  "label": "Portable Consent",
   "position": 11,
   "collapsible": true,
   "collapsed": true
