+++ package.json
@@ -30,7 +30,7 @@
     "@splidejs/react-splide": "^0.7.12",
     "@tailwindcss/line-clamp": "^0.4.2",
     "@xmtp/react-sdk": "^1.3.6",
-    "@xmtp/xmtp-js": "^8.0.2",
+    "@xmtp/xmtp-js": "^v11.3.0-beta.5",
     "buffer": "^6.0.3",
     "clsx": "^1.1.1",
     "docusaurus-plugin-plausible": "^0.0.5",
