+++ src/components/Widgets/FloatingInbox-text/ConversationContainer.js
@@ -1,0 +1,229 @@
+import React, { useState, useEffect } from "react";
+import { ethers } from "ethers";
+import { MessageContainer } from "./MessageContainer";
+import { ListConversations } from "./ListConversations";
+import { ListConversations as ListConversationsConsent } from "./ListConversations-consent";
+
+export const ConversationContainer = ({
+  client,
+  selectedConversation,
+  setSelectedConversation,
+  isContained = false,
+  isPWA = false,
+  isConsent = false,
+}) => {
+  const [searchTerm, setSearchTerm] = useState("");
+  const [peerAddress, setPeerAddress] = useState("");
+  const [message, setMessage] = useState("");
+  const [loading, setLoading] = useState(false);
+  const [loadingResolve, setLoadingResolve] = useState(false);
+  const [canMessage, setCanMessage] = useState(false);
+  const [conversationFound, setConversationFound] = useState(false);
+  const [createNew, setCreateNew] = useState(false);
+
+  const styles = {
+    conversations: {
+      height: "100%",
+    },
+    conversationList: {
+      padding: "0px",
+      marginTop: "0 !important",
+      listStyle: "none",
+      overflowY: "scroll",
+    },
+    conversationListItem: {
+      display: "flex",
+      marginTop: "0 !important",
+      justifyContent: "space-between",
+      alignItems: "center",
+      margin: "0px",
+      border: "0px",
+      borderBottom: "1px solid #e0e0e0",
+      cursor: "pointer",
+      backgroundColor: "#f0f0f0",
+      padding: "10px",
+      marginTop: "0px",
+      transition: "background-color 0.3s ease",
+    },
+    conversationDetails: {
+      display: "flex",
+      flexDirection: "column",
+      alignItems: "flex-start",
+      width: "75%",
+      marginLeft: "10px",
+      overflow: "hidden",
+    },
+    conversationName: {
+      fontSize: "16px",
+      fontWeight: "bold",
+    },
+    messagePreview: {
+      fontSize: "14px",
+      color: "#666",
+      whiteSpace: "nowrap",
+      overflow: "hidden",
+      textOverflow: "ellipsis",
+    },
+    conversationTimestamp: {
+      fontSize: "12px",
+      color: "#999",
+      width: "25%",
+      textAlign: "right",
+      display: "flex",
+      flexDirection: "column",
+      alignItems: "flex-end",
+      justifyContent: "space-between",
+    },
+    createNewButton: {
+      border: "1px",
+      padding: "5px",
+      borderRadius: "5px",
+      marginTop: "10px",
+    },
+    peerAddressInput: {
+      width: "100%",
+      padding: "10px",
+      boxSizing: "border-box",
+      border: "0px solid #ccc",
+    },
+  };
+
+  const selectConversation = async (conversation) => {
+    setSelectedConversation(conversation);
+  };
+
+  const isValidEthereumAddress = (address) => {
+    return /^0x[a-fA-F0-9]{40}$/.test(address);
+  };
+
+  const handleSearchChange = async (e) => {
+    setCreateNew(false);
+    setConversationFound(false);
+    setSearchTerm(e.target.value);
+    console.log("handleSearchChange", e.target.value);
+    setMessage("Searching...");
+    const addressInput = e.target.value;
+    const isEthDomain = /\.eth$/.test(addressInput);
+    let resolvedAddress = addressInput;
+    if (isEthDomain) {
+      setLoadingResolve(true);
+      try {
+        const provider = new ethers.providers.CloudflareProvider();
+        resolvedAddress = await provider.resolveName(resolvedAddress);
+      } catch (error) {
+        console.log(error);
+        setMessage("Error resolving address");
+        setCreateNew(false);
+      } finally {
+        setLoadingResolve(false);
+      }
+    }
+    console.log("resolvedAddress", resolvedAddress);
+    if (resolvedAddress && isValidEthereumAddress(resolvedAddress)) {
+      processEthereumAddress(resolvedAddress);
+      setSearchTerm(resolvedAddress); // <-- Add this line
+    } else {
+      setMessage("Invalid Ethereum address");
+      setPeerAddress(null);
+      setCreateNew(false);
+      //setCanMessage(false);
+    }
+  };
+
+  const processEthereumAddress = async (address) => {
+    setPeerAddress(address);
+    if (address === client.address) {
+      setMessage("No self messaging allowed");
+      setCreateNew(false);
+      // setCanMessage(false);
+    } else {
+      const canMessageStatus = await client?.canMessage(address);
+      if (canMessageStatus) {
+        setPeerAddress(address);
+        // setCanMessage(true);
+        setMessage("Address is on the network ✅");
+        setCreateNew(true);
+      } else {
+        //  setCanMessage(false);
+        setMessage("Address is not on the network ❌");
+        setCreateNew(false);
+      }
+    }
+  };
+  if (loading) {
+    return <div>Loading...</div>;
+  }
+  return (
+    <div style={styles.conversations}>
+      {!selectedConversation && (
+        <ul style={styles.conversationList}>
+          <input
+            type="text"
+            placeholder="Enter a 0x wallet or ENS address"
+            value={searchTerm}
+            onChange={handleSearchChange}
+            style={styles.peerAddressInput}
+          />
+          {loadingResolve && searchTerm && <small>Resolving address...</small>}
+          {isConsent ? (
+            <ListConversationsConsent
+              isPWA={isPWA}
+              client={client}
+              searchTerm={searchTerm}
+              selectConversation={setSelectedConversation}
+              onConversationFound={(state) => {
+                setConversationFound(state);
+                if (state == true) setCreateNew(false);
+              }}
+            />
+          ) : (
+            <ListConversations
+              isPWA={isPWA}
+              client={client}
+              searchTerm={searchTerm}
+              selectConversation={setSelectedConversation}
+              onConversationFound={(state) => {
+                setConversationFound(state);
+                if (state == true) setCreateNew(false);
+              }}
+            />
+          )}
+          {message && conversationFound !== true && <small>{message}</small>}
+          {peerAddress && canMessage && (
+            <button
+              style={styles.createNewButton}
+              onClick={() => {
+                setSelectedConversation({ messages: [] });
+              }}>
+              Create new conversation
+            </button>
+          )}
+        </ul>
+      )}
+      {selectedConversation && (
+        <MessageContainer
+          client={client}
+          isContained={isContained}
+          conversation={selectedConversation}
+          searchTerm={searchTerm}
+          isConsent={isConsent}
+          selectConversation={selectConversation}
+        />
+      )}
+    </div>
+  );
+};
+
+const getRelativeTimeLabel = (dateString) => {
+  const diff = new Date() - new Date(dateString);
+  const diffMinutes = Math.floor(diff / 1000 / 60);
+  const diffHours = Math.floor(diff / 1000 / 60 / 60);
+  const diffDays = Math.floor(diff / 1000 / 60 / 60 / 24);
+  const diffWeeks = Math.floor(diff / 1000 / 60 / 60 / 24 / 7);
+
+  if (diffMinutes < 60)
+    return `${diffMinutes} minute${diffMinutes > 1 ? "s" : ""} ago`;
+  if (diffHours < 24) return `${diffHours} hour${diffHours > 1 ? "s" : ""} ago`;
+  if (diffDays < 7) return `${diffDays} day${diffDays > 1 ? "s" : ""} ago`;
+  return `${diffWeeks} week${diffWeeks > 1 ? "s" : ""} ago`;
+};
